Add files via upload
This bot will help you collect the viewers comments from any YouTube video of your choosing. All you need to do is provide a YouTube video link and the script will fetch the top 65 comments for you, as well as the authors\' names, the number of likes for each comment and the date each comment was published. The bot will then save these details to a csv file (excel / spreadsheet type of file). Please use responsibly.